Create README.md
The objective of this project, is to demonstratre Python and object-oriented programming skills to build a program that plays the game of Rock Paper Scissors. In this project, I have builts classes that represent the game and its players. I have written computer players that follow various different strategies, as well as a human player class that lets a human play the game against the computer. Following is a summary of the classes: 

1. A RandomPlayer has been created. It chooses its move at random and returns one of 'rock', 'paper', or 'scissors' at random.
2. The game class displays the outcome of each round, and keeps score for both players. The game class can handle ties — when both players make the same move!
3. A HumanPlayer subclass, whose move method asks the human user what move to make. 
Set the program to play a game between HumanPlayer and RandomPlayer.
4. A ReflectPlayer class that remembers what move the opponent played last round, and plays that move this round. 
5. A CyclePlayer class that remembers what move it played last round, and cycles through the different moves. 

The human player might sometimes make typos. If they enter roxk instead of rock, the HumanPlayer code should let them try again. 

The game let the user choose the number of rounds at the beginning of the game and at the end of the game, have it print out which player won, and what the final scores are.

check_inputs.py: file contains the code to validate the user input
game_rules.py: contains the rules of the game- Paper beats rock; rock beats scissors; scissors beat paper.
player.py: contains all the player class
game.py: uses all the files and let the user play the game.